New PerProcessThrottler
Will requeue jobs immediately when a processes' rate limit has been reached.
This is instead of placing the job in the scheduled queue.

This gives the opportunity for other processes that haven't reached their
rate limit to process the job.

E.g., if you want to load balance (via rate limiting) for a set of
distributed workers.