Perfoming another perform_async instead of pushing direclty to queue
Looks like I need to be higher up in the chain of execution to get
at the raw data to push back on the queue.

The downside to this is that jobs that are popped but not executed
will still show up as being processed in the stats counter.

Need to fix that...